* commit '4af9cfb292f33d06bff01467205e324c563d5af8':
  doxygen toegevoegd
  update filelogger, log per day, resultlogger
  update configuraties and fix some overflow :)
  add break and repair calculations

+++ src/libs/models/Machine.cpp
@@ -87,8 +87,8 @@ uint16_t Machine::getMeanTimeBetweenFailureInHours() const {
   return meanTimeBetweenFailureInHours;
 }
 
-uint32_t Machine::getMeanTimeBetweenFailureInMillis() const {
-  return 3600000u * meanTimeBetweenFailureInHours;
+uint64_t Machine::getMeanTimeBetweenFailureInMillis() const {
+  return 3600000lu * meanTimeBetweenFailureInHours;
 }
 
 uint16_t Machine::getReparationTimeInMinutes() const {
